Move functions under utils/nested
