+++ simpeg_drivers/components/factories/misfit_factory.py
@@ -11,27 +11,19 @@
 
 from __future__ import annotations
 
-from concurrent.futures import ProcessPoolExecutor, as_completed
-from copy import copy
-from pathlib import Path
 from typing import TYPE_CHECKING
 
 import numpy as np
-from discretize import TensorMesh, TreeMesh
-from geoh5py.objects import Octree
-from simpeg import data, data_misfit, maps, meta, objective_function
+from simpeg import objective_function
 from simpeg.simulation import BaseSimulation
 
 from simpeg_drivers.components.factories.simpeg_factory import SimPEGFactory
-from simpeg_drivers.components.factories.simulation_factory import SimulationFactory
-from simpeg_drivers.utils.utils import create_nested_mesh
+from simpeg_drivers.utils.nested import create_nested_misfit
 
 
 if TYPE_CHECKING:
     from geoapps_utils.driver.params import BaseParams
 
-    from simpeg_drivers.components.data import InversionData
-    from simpeg_drivers.components.meshes import InversionMesh
     from simpeg_drivers.options import BaseOptions
 
 
@@ -74,6 +66,8 @@ class MisfitFactory(SimPEGFactory):
         tile_count = 0
         data_count = 0
         misfit_count = 0
+
+        # TODO bring back on GEOPY-2182
         # with ProcessPoolExecutor() as executor:
         for local_index in tiles:
             if len(local_index) == 0:
@@ -110,273 +104,3 @@ class MisfitFactory(SimPEGFactory):
     def assemble_keyword_arguments(self, **_):
         """Implementation of abstract method from SimPEGFactory."""
         return {}
-
-
-def create_nested_misfit(
-    simulation,
-    local_index,
-    channels,
-    tile_count,
-    # data_count,
-    n_split,
-    padding_cells,
-    inversion_type,
-    forward_only,
-):
-    local_sim, _ = create_nested_simulation(
-        simulation,
-        None,
-        local_index,
-        channel=None,
-        tile_id=tile_count,
-        padding_cells=padding_cells,
-    )
-
-    local_mesh = getattr(local_sim, "mesh", None)
-    sorting = []
-    local_misfits = []
-    for count, channel in enumerate(channels):
-        for split_ind in np.array_split(local_index, n_split[count]):
-            local_sim, mapping = create_nested_simulation(
-                simulation,
-                local_mesh,
-                split_ind,
-                channel=channel,
-                tile_id=tile_count,
-                padding_cells=padding_cells,
-            )
-
-            if count == 0:
-                sorting.append(split_ind)
-
-            meta_simulation = meta.MetaSimulation(
-                simulations=[local_sim], mappings=[mapping]
-            )
-
-            local_data = data.Data(local_sim.survey)
-            lmisfit = data_misfit.L2DataMisfit(local_data, meta_simulation)
-            if not forward_only:
-                local_data.dobs = local_sim.survey.dobs
-                local_data.standard_deviation = local_sim.survey.std
-                name = inversion_type
-                name += f": Tile {tile_count + 1}"
-                if len(channels) > 1:
-                    name += f": Channel {channel}"
-
-                lmisfit.name = f"{name}"
-
-            local_misfits.append(lmisfit)
-
-            tile_count += 1
-
-    return local_misfits, sorting
-
-
-def create_nested_simulation(
-    simulation: BaseSimulation,
-    local_mesh: TreeMesh | None,
-    indices: np.ndarray,
-    *,
-    channel: int | None = None,
-    tile_id: int | None = None,
-    padding_cells=100,
-):
-    """
-    Generate a survey, mesh and simulation based on indices.
-
-    :param inversion_data: InversionData object.
-    :param mesh: Octree mesh.
-    :param active_cells: Active cell model.
-    :param indices: Indices of receivers belonging to the tile.
-    :param channel: Channel number for frequency or time channels.
-    :param tile_id: Tile id stored on the simulation.
-    :param padding_cells: Number of padding cells around the local survey.
-    """
-    local_survey = create_nested_survey(
-        simulation.survey, indices=indices, channel=channel
-    )
-
-    if local_mesh is None:
-        local_mesh = create_nested_mesh(
-            local_survey,
-            simulation.mesh,
-            minimum_level=3,
-            padding_cells=padding_cells,
-        )
-
-    mapping = maps.TileMap(
-        simulation.mesh,
-        simulation.active_cells,
-        local_mesh,
-        enforce_active=True,
-        components=3 if getattr(simulation, "model_type", None) == "vector" else 1,
-    )
-
-    kwargs = {"survey": local_survey}
-
-    n_actives = int(mapping.local_active.sum())
-    if getattr(simulation, "_chiMap", None) is not None:
-        if simulation.model_type == "vector":
-            kwargs["chiMap"] = maps.IdentityMap(nP=n_actives * 3)
-            kwargs["model_type"] = "vector"
-        else:
-            kwargs["chiMap"] = maps.IdentityMap(nP=n_actives)
-
-        kwargs["active_cells"] = mapping.local_active
-        kwargs["sensitivity_path"] = (
-            Path(simulation.sensitivity_path).parent / f"Tile{tile_id}.zarr"
-        )
-
-    if getattr(simulation, "_rhoMap", None) is not None:
-        kwargs["rhoMap"] = maps.IdentityMap(nP=n_actives)
-        kwargs["active_cells"] = mapping.local_active
-        kwargs["sensitivity_path"] = (
-            Path(simulation.sensitivity_path).parent / f"Tile{tile_id}.zarr"
-        )
-
-    if getattr(simulation, "_sigmaMap", None) is not None:
-        kwargs["sigmaMap"] = maps.ExpMap(local_mesh) * maps.InjectActiveCells(
-            local_mesh, mapping.local_active, value_inactive=np.log(1e-8)
-        )
-
-    if getattr(simulation, "_etaMap", None) is not None:
-        kwargs["etaMap"] = maps.InjectActiveCells(
-            local_mesh, mapping.local_active, value_inactive=0
-        )
-        proj = maps.InjectActiveCells(
-            local_mesh,
-            mapping.local_active,
-            value_inactive=1e-8,
-        )
-        kwargs["sigma"] = proj * mapping * simulation.sigma
-
-    for key in [
-        "max_chunk_sizestore_sensitivities",
-        "solver",
-        "t0",
-        "time_steps",
-        "thicknesses",
-    ]:
-        if hasattr(simulation, key):
-            kwargs[key] = getattr(simulation, key)
-
-    local_sim = type(simulation)(local_mesh, **kwargs)
-
-    # TODO bring back
-    # inv_type = inversion_data.params.inversion_type
-    # if inv_type in ["fdem", "tdem"]:
-    #     compute_em_projections(inversion_data, local_sim)
-    # elif ("current" in inv_type or "polarization" in inv_type) and (
-    #     "2d" not in inv_type or "pseudo" in inv_type
-    # ):
-    #     compute_dc_projections(inversion_data, local_sim, indices)
-    return local_sim, mapping
-
-
-def create_nested_survey(survey, indices, channel=None):
-    """
-    Extract source and receivers belonging to the indices.
-    """
-    sources = []
-    location_count = 0
-    for src in survey.source_list or [survey.source_field]:
-        if channel is not None and getattr(src, "frequency", None) != channel:
-            continue
-
-        # Extract the indices of the receivers that belong to this source
-        locations = src.receiver_list[0].locations
-        if isinstance(locations, tuple | list):  # For MT survey
-            n_data = locations[0].shape[0]
-        else:
-            n_data = locations.shape[0]
-
-        rx_indices = np.arange(n_data) + location_count
-
-        _, intersect, _ = np.intersect1d(rx_indices, indices, return_indices=True)
-        location_count += n_data
-
-        if len(intersect) == 0:
-            continue
-
-        receivers = []
-        for rx in src.receiver_list:
-            # intersect = set(rx.local_index).intersection(indices)
-            new_rx = copy(rx)
-
-            if isinstance(rx.locations, tuple | list):  # For MT and DC surveys
-                new_rx.locations = tuple(loc[intersect] for loc in rx.locations)
-            else:
-                new_rx.locations = rx.locations[intersect]
-
-            receivers.append(new_rx)
-
-        if any(receivers):
-            new_src = copy(src)
-            new_src.receiver_list = receivers
-            sources.append(new_src)
-
-    if hasattr(survey, "source_field"):
-        new_survey = type(survey)(sources[0])
-    else:
-        new_survey = type(survey)(sources)
-
-    if hasattr(survey, "dobs") and survey.dobs is not None:
-        n_channels = len(np.unique(survey.ordering[:, 0]))
-        n_comps = len(np.unique(survey.ordering[:, 1]))
-        order = "C" if hasattr(survey, "frequencies") else "F"
-        data_slice = survey.dobs.reshape((n_channels, n_comps, -1), order=order)[
-            :, :, indices
-        ]
-        uncert_slice = survey.std.reshape((n_channels, n_comps, -1), order=order)[
-            :, :, indices
-        ]
-
-        # For FEM surveys only
-        if channel is not None:
-            ind = np.where(np.asarray(survey.frequencies) == channel)[0]
-            data_slice = data_slice[ind, :, :]
-            uncert_slice = uncert_slice[ind, :, :]
-
-        new_survey.dobs = data_slice.flatten(order=order)
-        new_survey.std = uncert_slice.flatten(order=order)
-
-    return new_survey
-
-
-def compute_em_projections(inversion_data, simulation):
-    """
-    Pre-compute projections for the receivers for efficiency.
-    """
-    rx_locs = inversion_data.entity.vertices
-    projections = {}
-    for component in "xyz":
-        projections[component] = simulation.mesh.get_interpolation_matrix(
-            rx_locs, "faces_" + component[0]
-        )
-
-    for source in simulation.survey.source_list:
-        for receiver in source.receiver_list:
-            projection = 0.0
-            for orientation, comp in zip(receiver.orientation, "xyz", strict=True):
-                if orientation == 0:
-                    continue
-                projection += orientation * projections[comp][receiver.local_index, :]
-            receiver.spatialP = projection
-
-
-def compute_dc_projections(inversion_data, simulation, indices):
-    """
-    Pre-compute projections for the receivers for efficiency.
-    """
-    rx_locs = inversion_data.entity.vertices
-    mn_pairs = inversion_data.entity.cells
-    projection = simulation.mesh.get_interpolation_matrix(rx_locs, "nodes")
-
-    for source, ind in zip(simulation.survey.source_list, indices, strict=True):
-        proj_mn = projection[mn_pairs[ind, 0], :]
-
-        # Check if dipole receiver
-        if not np.all(mn_pairs[ind, 0] == mn_pairs[ind, 1]):
-            proj_mn -= projection[mn_pairs[ind, 1], :]
-
-        source.receiver_list[0].spatialP = proj_mn  # pylint: disable=protected-access
